refactor: App파일 분리 -> MovieList 컴포넌트 생성
