+++ docs/REFACTOR.md
@@ -7,6 +7,9 @@
 
 - [x] css 분리, 리팩토링 (공통, 컴포넌트, 변수화)
 - [x] ErrorPage 파일명에 대한 고려, 매직넘버 처리
-- [ ] 코드의 일관성
+- [x] App 파일 분리 -> MovieList 컴포넌트로
+- [ ] API 호출 분리
+- [ ] 스켈레톤 대량 추가시, DOM에는 한번만 접근하는 방식 -> 성능 이점
 
+- [ ] 코드의 일관성, 중복 로직 제거
 - [ ] 무한스크롤 구현하며 API에서 몇개의 데이터가 올지 예상, Skeleton 구현
